+++ changelog.md
@@ -7,6 +7,7 @@
 - Fixed Hide if not relevant General not working
 - Fixed Hide if not relevant Deployables not working on Power Orbs
 - Fixed Hide if not relevant Trophy fishing not working
+- Fixed Golden Fish Timer to now properly take into account rod casting.
 # V0.2.6
 - Added new github version alert.
 # V0.2.5